+++ lib/blocs/settings/speech/speech_settings_cubit.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 import 'package:bloc/bloc.dart';
 import 'package:flutter_download_manager/flutter_download_manager.dart';
 import 'package:lotti/blocs/settings/speech/speech_settings_state.dart';
+import 'package:lotti/database/settings_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/asr_service.dart';
 import 'package:lotti/utils/file_utils.dart';
@@ -26,12 +27,22 @@ class SpeechSettingsCubit extends Cubit<SpeechSettingsState> {
       _downloadProgress[model] = 0;
     }
     detectDownloadedModels();
+    loadSelectedModel();
   }
 
   final AsrService _asrService = getIt<AsrService>();
   Map<String, double> _downloadProgress = <String, double>{};
   String _selectedModel = '';
   late final DownloadManager _downloadManager;
+
+  Future<void> loadSelectedModel() async {
+    final selectedModel = await getIt<SettingsDb>().itemByKey(whisperModelKey);
+
+    if (selectedModel != null) {
+      _selectedModel = selectedModel;
+      emitState();
+    }
+  }
 
   Future<void> detectDownloadedModels() async {
     final docDir = await findDocumentsDirectory();
@@ -51,9 +62,15 @@ class SpeechSettingsCubit extends Cubit<SpeechSettingsState> {
     emitState();
   }
 
-  void selectModel(String selectedModel) {
+  Future<void> selectModel(String selectedModel) async {
     _selectedModel = selectedModel;
     _asrService.model = selectedModel;
+
+    await getIt<SettingsDb>().saveSettingsItem(
+      whisperModelKey,
+      selectedModel,
+    );
+
     emitState();
   }
 
@@ -72,7 +89,13 @@ class SpeechSettingsCubit extends Cubit<SpeechSettingsState> {
     }
 
     task.progress.addListener(() {
-      _downloadProgress[model] = task.progress.value;
+      final value = task.progress.value;
+      _downloadProgress[model] = value;
+
+      if (value == 1.0 && _selectedModel.isEmpty) {
+        _selectedModel = model;
+      }
+
       emitState();
     });
 
@@ -82,6 +105,10 @@ class SpeechSettingsCubit extends Cubit<SpeechSettingsState> {
 
   Future<void> deleteModel(String model) async {
     _downloadProgress[model] = 0;
+    if (_selectedModel == model) {
+      _selectedModel = '';
+    }
+
     emitState();
     final fileName = 'ggml-$model.bin';
     final docDir = await findDocumentsDirectory();